Fix empty reference value selection modals
If a content type has both a Reference field and a reference_of field
(like a 'Category' content type with a 'Products' reference field and
a 'Child Caregories' field), the condition (parent.content = <id>)
used in the uniteViewDataFetcher object for the reference_of field
will "leak" and remain set if any other uniteViewDataFetcher objects
are created. In this example, this means any new rows added to the
'Products' reference collection will have empty modals when the plus
button is clicked, as they're they're fetching products that are
already children of the current category. This is becaue they have the
filter from the 'Child Categories' field applied when they do a
GraphQL query to fetch the options to show.

This occurs becase the uniteViewDataFetcher create function returns a
shallow clone of itself, while the filter conditions are an object
reference that isn't cloned. To work around the issue, this method now
does a deep clone initially, modifies it and returns it.