Fix moving tree items to lower levels
Previously, it wasn't possible to move tree view items to be children
of other items if the other item didn't already have children. This
was because the empty tree-view-children node wasn't rendered for
items without any children, meaning that potential new children had
nowhere to be moved to.

Also add the `uk-sortable` class to tree-view-children nodes. The
uikit3-nestable plugin uses this to determine which nodes items can be
nested under, so moving items under other items didn't work properly
without it.