Add rows_per_page setting to table and tree views
This setting controls how many rows are rendered per page. Previously
this was hard-coded to 10, which caused problems for tree views - it
isn't possible to drag items from one page to another, so all the top-
level items need to fit on one page.

The `rows_per_page` setting is optional, and will default to 10 if not
specified. If specified, it must be an integer - floats and strings
are not accepted. An example of config using the setting:

```
{
  "content_types": [
    {
      "title": "Categories",
      "identifier": "category",
      "fields": [ ... ],
      "views": [
        {
          "title": "All",
          "identifier": "all",
          "type": "table",
          "settings": {
            "fields": { ... },
            "rows_per_page": 20
          }
        }
      ]
    }
  ]
}
```