Move FieldTypeManager dependency into factories
Move the FieldTypeManager dependency into the factory services, rather
than requiring it to be passed in to their `create` methods. The
FieldTypeManager service was no longer used by TableViewType, so it
was removed from that class.

Also remove the `createConfig` method from TreeViewType and
GridViewType, so they use the version inherited from TableViewType
(and so actually use the factory to create the config object).